BSP-116: Types & Constructors fixes & commands moved to single folder
