+++ src/commands/authnCommand.ts
@@ -7,8 +7,8 @@ export default class AuthnCommand implements AuthnInterface {
     readonly username: string;
     readonly password: string;
 
-    constructor(data: unknown) {
-        const { username, password } = data as AuthnInterface;
+    constructor(data: Record<string, unknown> | AuthnInterface) {
+        const { username, password } = data;
 
         if (typeof username !== "string") {
             throw new Error("INVALID_USERNAME");
