+++ src/commands/createSessionCommand.ts
@@ -5,8 +5,8 @@ interface CreateSessionInterface {
 export default class CreateSessionCommand implements CreateSessionInterface {
     readonly sessionToken: string;
 
-    constructor(data: unknown) {
-        const { sessionToken } = data as CreateSessionInterface;
+    constructor(data: Record<string, unknown> | CreateSessionInterface) {
+        const { sessionToken } = data;
 
         if (typeof sessionToken !== "string") {
             throw new Error("INVALID_SESSION_TOKEN");
