+++ src/commands/userinfoCommand.ts
@@ -5,8 +5,8 @@ interface UserinfoInterface {
 export default class UserinfoCommand implements UserinfoInterface {
     readonly bearerToken: string;
 
-    constructor(data: unknown) {
-        const { bearerToken } = data as UserinfoInterface;
+    constructor(data: Record<string, unknown> | UserinfoInterface) {
+        const { bearerToken } = data;
 
         if (typeof bearerToken !== "string") {
             throw new Error("INVALID_BEARER");
